Moved the floor stuff to entities/floor.rs - JIRA-47 - successfully finishing the split of all the possible things from main.rs to separate files!
